+++ tests/test_status_map.py
@@ -93,6 +93,8 @@ def test_status_add_previous_list():
 
 
 def test_status_map(status_map, transitions):
+    assert has_same_elements(status_map.statuses, ('pending', 'processing', 'approved', 'rejected', 'processed'))
+    assert str(status_map.statuses) in repr(status_map)
     assert status_map._transitions == transitions
     assert status_map._transitions is not transitions
     assert status_map._statuses